+++ xarray/tests/__init__.py
@@ -2,13 +2,13 @@ import importlib
 import platform
 import warnings
 from contextlib import contextmanager
-from distutils import version
 from unittest import mock  # noqa: F401
 
 import numpy as np
 import pandas as pd
 import pytest
 from numpy.testing import assert_array_equal  # noqa: F401
+from packaging.version import Version
 from pandas.testing import assert_frame_equal  # noqa: F401
 
 import xarray.testing
@@ -45,19 +45,12 @@ def _importorskip(modname, minversion=None):
         mod = importlib.import_module(modname)
         has = True
         if minversion is not None:
-            if LooseVersion(mod.__version__) < LooseVersion(minversion):
+            if Version(mod.__version__) < Version(minversion):
                 raise ImportError("Minimum version not satisfied")
     except ImportError:
         has = False
     func = pytest.mark.skipif(not has, reason=f"requires {modname}")
     return has, func
-
-
-def LooseVersion(vstring):
-    # Our development version is something like '0.10.9+aac7bfc'
-    # This function just ignored the git commit id.
-    vstring = vstring.split("+")[0]
-    return version.LooseVersion(vstring)
 
 
 has_matplotlib, requires_matplotlib = _importorskip("matplotlib")